+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 
+ * Improve exception message for missing value [#131](https://github.com/premailer/css_parser/pull/131)
  * `:rule_set_exceptions` option added [#132](https://github.com/premailer/css_parser/pull/132)
 
 ### Version 1.11.0